+++ lib/game_ar.h
@@ -1,7 +1,10 @@
 #ifndef _GAME_AR_H_
 #define _GAME_AR_H_
 
+#include <game.h>
+#include <function_interface.h>
 game new_game_ar( int nb_pieces, piece *pieces);
 bool game_over_ar(cgame g);
+game getGameFromIdAR(char* id);
 
 #endif
