+++ lib/interface_txt.h
@@ -4,7 +4,6 @@
 #include "function_interface.h"
 
 game getGameFromIdAR(char*  id);
-void getIdFromGameAR(game g, char* id);
 void getIdFromGame(game g, char* id);
 void draw_interface(game g, char* id);
 void getHelp(int input, bool* done);
